requirements.txt
Flask==2.1.0
gunicorn==20.0.0
json5==0.8.5
jsonschema==3.0.2
pandas==1.3.5
spotipy==2.4.4
numpy==1.17.3
mpld3==0.3
scipy==1.6.0
vaderSentiment==3.2.1
lyricsgenius==3.0.0
requests==2.22.0
wordcloud==1.6.0
plotly==4.2.1
Flask-PyMongo==2.3.0
bs4==0.0.1
dnspython==1.16.0
gensim==3.8.3
spacy==3.3.1
nltk==3.4.5
NRCLex==2.0.2
textblob==0.15.3
python-decouple==3.3
python-dotenv==0.13.0
lexical-diversity==0.1.1
openai==0.20.0
urllib3==1.26.6
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#egg=en_core_web_sm
